Unify szr_stats
